feat: rename controlled-input to inputs packages and move checkbox to this new package
