+++ packages/components/inputs/README.mdx
@@ -1,3 +1,3 @@
-# @contentful/f36-controlled-input
+# @contentful/f36-inputs
 
 This package is part of [forma-36](https://github.com/contentful/forma-36). See the repo for more details.
