+++ packages/components/inputs/src/checkbox/Checkbox.tsx
@@ -1,9 +1,6 @@
 import React from 'react';
 
-import {
-  ControlledInput,
-  ControlledInputProps,
-} from '@contentful/f36-controlled-input';
+import { ControlledInput, ControlledInputProps } from '..';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 export interface CheckboxProps extends ControlledInputProps {}
