+++ packages/components/inputs/src/index.ts
@@ -1,0 +1,2 @@
+export * from './controlled-input';
+export * from './checkbox';
