--- packages/forma-36-react-components/src/components/Checkbox/README.mdx
@@ -1,10 +1,0 @@
----
-title: 'Checkbox'
-type: 'component'
-status: 'stable'
-slug: /components/checkbox/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/forma-36-react-components/src/components/Checkbox'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/components-checkbox--overview'
----
-
-Detailed documentation will be provided soon. For now, please have a look into the storybook for implementation details.
