+++ packages/forma-36-react-components/src/components/CheckboxField/README.mdx
@@ -3,8 +3,8 @@ title: 'CheckboxField'
 type: 'component'
 status: 'stable'
 slug: /components/checkbox-field/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/forma-36-react-components/src/components/Checkbox'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/components-checkbox--default'
+github: 'https://github.com/contentful/forma-36/tree/master/packages/forma-36-react-components/src/components/CheckboxField'
+storybook: 'https://f36-storybook.contentful.com/?path=/story/form-elements-checkboxfield--basic'
 ---
 
 Use CheckboxField to give users the possibility to select multiple elements, zero, one or more items. This component can be used for example in forms or terms and conditions. It contains a checkbox with its label. For a single choice, please consider using [RadioButtonField component](https://f36.contentful.com/components/radio-button-field/).
