+++ packages/forma-36-website/gatsby-config.js
@@ -191,7 +191,7 @@ module.exports = {
             menuLinks: [
               {
                 name: 'Checkbox',
-                link: '/components/checkbox/',
+                link: '/components/inputs/src/checkbox/',
               },
               {
                 name: 'CheckboxField',
